Changed the duration of stay to show 0 or N > 0. Fixed the arrival and checkout date div to show '----' instead of 'Invalid Date'. Change the quote inside lines 144 and 150 to '' if we want to have nothing show. Might look more Professional?
Added unmbers to the gallery pagination dots and made them square-ish so the numbers are easier to read.

Changed the room h1 to be smaller and added some padding. Renove the small heading 'roooms' under the title of the page as we only have rooms at the moment so its redundant.